refactor: 이벤트 event 패키지로 이동
