+++ app/src/main/java/kitchenpos/application/order/OrderService.java
@@ -13,8 +13,8 @@ import kitchenpos.domain.order.OrderLineItem;
 import kitchenpos.domain.order.OrderLineItems;
 import kitchenpos.domain.order.OrderRepository;
 import kitchenpos.domain.order.OrderStatus;
-import kitchenpos.domain.order.OrderedEvent;
-import kitchenpos.domain.order.OrderedMenuEvent;
+import kitchenpos.event.OrderedEvent;
+import kitchenpos.event.OrderedMenuEvent;
 import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
