+++ app/src/main/java/kitchenpos/application/table/OrderedEventListener.java
@@ -2,9 +2,9 @@ package kitchenpos.application.table;
 
 import static kitchenpos.exception.order.OrderExceptionType.ORDER_TABLE_CAN_NOT_EMPTY;
 
-import kitchenpos.domain.order.OrderedEvent;
 import kitchenpos.domain.table.OrderTable;
 import kitchenpos.domain.table.OrderTableRepository;
+import kitchenpos.event.OrderedEvent;
 import kitchenpos.exception.order.OrderException;
 import org.springframework.context.event.EventListener;
 import org.springframework.stereotype.Component;
